requirements.txt
flask
requests
gspread
google-auth
python-telegram-bot==20.7
google-generativeai
pydantic==1.10.12
typing-extensions==4.8.0
aiohttp
